Stats Pre-aggregation for performance critical APIs

+++ src/MiningCore/Blockchain/Bitcoin/BitcoinPayoutHandler.cs
@@ -189,9 +189,9 @@ namespace MiningCore.Blockchain.Bitcoin
             return result.ToArray();
         }
 
-        public virtual Task CalculateBlockEffortAsync(Block block, ulong accumulatedBlockShareDiff)
+        public virtual Task CalculateBlockEffortAsync(Block block, double accumulatedBlockShareDiff)
         {
-            block.Effort = (double) accumulatedBlockShareDiff / block.NetworkDifficulty;
+            block.Effort = accumulatedBlockShareDiff / block.NetworkDifficulty;
 
             return Task.FromResult(true);
         }
