+++ src/MiningCore/Blockchain/Monero/MoneroPayoutHandler.cs
@@ -393,9 +393,9 @@ namespace MiningCore.Blockchain.Monero
             return result.ToArray();
         }
 
-        public Task CalculateBlockEffortAsync(Block block, ulong accumulatedBlockShareDiff)
+        public Task CalculateBlockEffortAsync(Block block, double accumulatedBlockShareDiff)
         {
-            block.Effort = (double) accumulatedBlockShareDiff / block.NetworkDifficulty;
+            block.Effort = accumulatedBlockShareDiff / block.NetworkDifficulty;
 
             return Task.FromResult(true);
         }
